Did an extensive overhaul of the code structure, getting the proto stuff in its own directory and making a proper module out of it.  Additionally, certificates and leys were also put into their own directory.
Then refactored the code so tht the logging
mechanism was standardized in one place, and
ensured that the server was using it instead of
just print statements.

Lastly, refactored the configuration elements into
its own clase and used inheritance in order to
sort out the differences between the client and
server code options.  Now the code uses
dependency injection (somewhat) in order to handle
config options which keeps the client and server
code a lot more tidy.

A few other minor fixes and changes.